actually use trace instead of debug and fix some test bugs

+++ svalin_rpc/src/test/aucpace_test_command.rs
@@ -48,7 +48,7 @@ impl TakeableCommandHandler for AucPaceTestCommandHandler {
 
             let mut session_ready = Session::new(Box::new(transport), Peer::Anonymous);
 
-            let ping: u64 = session_ready.read_object().await?;
+            let ping: u128 = session_ready.read_object().await?;
             session_ready.write_object(&ping).await?;
 
             *session = Some(session_ready);
